Type of dimension<T>::value should be int

+++ gtsam/geometry/Cal3Bundler.h
@@ -174,7 +174,7 @@ struct is_manifold<Cal3Bundler> : public std::true_type {
 };
 
 template<>
-struct dimension<Cal3Bundler> : public std::integral_constant<size_t, 3> {
+struct dimension<Cal3Bundler> : public std::integral_constant<int, 3> {
 };
 
 } // namespace gtsam
